Refactor: productState로직 분리
